DXC_DCG2026/requirements.txt
# ============================================
# requirements.txt — LIK Insurance Analyst
# Versions flexibles (>=) pour compatibilité
# ============================================

# ------------------------------
# INTERFACE & APPLICATION
# ------------------------------

streamlit>=1.28.0
# Framework web pour dashboards interactifs (upload, visualisations, IA)

python-dotenv>=1.0.0
# Chargement sécurisé des variables d’environnement (.env)

# ------------------------------
# TRAITEMENT & STRUCTURE DES DONNÉES
# ------------------------------

pandas>=2.0.0
# Manipulation des données tabulaires (DataFrame, nettoyage, agrégation)

numpy>=1.24.0
# Calcul numérique de base (fondation de pandas et scikit-learn)

scipy>=1.10.0
# Statistiques avancées, distributions, calcul scientifique

pyarrow>=12.0.0
# Lecture/écriture performante de gros fichiers (CSV, Parquet)

openpyxl>=3.1.0
# Lecture et écriture des fichiers Excel (.xlsx)

# ------------------------------
# VISUALISATION
# ------------------------------

plotly>=5.0.0
# Visualisations interactives (graphiques dynamiques)

matplotlib>=3.7.0
# Graphiques statiques et exports

seaborn>=0.12.0
# Visualisations statistiques avancées (EDA)

# ------------------------------
# MACHINE LEARNING & SCORING
# ------------------------------

scikit-learn>=1.3.0
# Algorithmes ML, feature selection, pipelines

imbalanced-learn>=0.10.0
# Gestion du déséquilibre de classes (SMOTE, undersampling)

joblib>=1.3.0
# Sauvegarde et chargement de modèles ML

# ------------------------------
# INTELLIGENCE ARTIFICIELLE & NLP
# ------------------------------

openai>=1.3.0
# Accès à l’API OpenAI (assistant IA, NLQ)

langchain>=0.0.300
# Orchestration des appels LLM (chaînes, prompts)

langchain-openai>=0.0.5
# Connecteur officiel OpenAI pour LangChain

# ------------------------------
# GÉNÉRATION DE RAPPORTS & DOCUMENTS
# ------------------------------

reportlab>=4.0.0
# Génération de rapports PDF

python-docx>=0.8.11
# Génération de documents Word (.docx)

markdown>=3.4.0
# Rendu et conversion Markdown

jinja2>=3.1.0
# Templates HTML / rapports dynamiques

PyPDF2>=3.0.0
# Manipulation de fichiers PDF

fpdf2>=2.7.0
# Génération PDF simple et rapide

# ------------------------------
# STATISTIQUES AVANCÉES (OPTIONNEL)
# ------------------------------

statsmodels>=0.13.0
# Modèles statistiques (régressions, séries temporelles)

pingouin>=0.5.0
# Tests statistiques prêts à l’emploi (ANOVA, corrélations)

# ------------------------------
# UTILITAIRES GÉNÉRAUX
# ------------------------------

python-dateutil>=2.8.0
# Manipulation avancée des dates/heures

pytz>=2023.0
# Gestion des fuseaux horaires

tqdm>=4.64.0
# Barres de progression pour traitements longs

loguru>=0.6.0
# Logging lisible et structuré

prophet

torch
requests>=2.31.0
# Requêtes HTTP (APIs externes)

# ------------------------------
# MODÈLES AVANCÉS (OPTIONNELS)
# ------------------------------

catboost >= 1.2.0


xgboost>=1.7.0
# Modèle de boosting performant pour le scoring

# catboost>=1.2.0
# Modèle performant mais souvent instable à installer sous Windows (optionnel)
# requirements.txt - Mise à jour pour correspondre à setup.py

# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
scipy>=1.10.0
joblib>=1.3.0
openai>=1.3.0
langchain>=0.0.300
langchain-openai>=0.0.5
python-dateutil>=2.8.0
tqdm>=4.64.0
requests>=2.31.0

# Optional dependencies (commented by default)
# xgboost>=1.7.0
# lightgbm>=4.0.0
# catboost>=1.2.0
# statsmodels>=0.14.0
# prophet>=1.1.0
# shap>=0.42.0
# imbalanced-learn>=0.11.0
# reportlab>=4.0.0
# python-docx>=0.8.11
# markdown>=3.4.0
# jinja2>=3.1.0
# PyPDF2>=3.0.0
# fpdf2>=2.7.0
# pingouin>=0.5.0
# loguru>=0.6.0